fix(epd): widen the rendezvous bootstrap_room to 63 bits (TokenSpeed path)
The gateway mints bootstrap_room as a random i32 in [0, 2^31) per image (E->P
leg) and per request (P->D leg) with NO in-flight dedup. The collision domain
is per rendezvous endpoint (a prefill worker's embedding receiver, a decode
worker's KV receiver): with tens-to-hundreds of img/s system-wide, room
lifetimes of seconds-to-minutes under queueing, and thousands of rooms live
concurrently per worker, the birthday bound puts an expected collision within
a few days of sustained load. A collision silently cross-wires two requests'
embedding/KV rendezvous (one image's embedding written onto another's room) or
leaks an RDMA arena slot -- a silent correctness fault, not a crash. Multiple
independent gateways (each minting in the same 2^31 space) only worsen it, and
that is exactly the double-gateway scale-out direction.

Widen the TokenSpeed rendezvous id to int64 end to end: the random mint draws
from [0, 2^63), so the birthday collision rate is negligible even with many
gateways and thousands of live rooms. The wire path was already width-agnostic
(every leg encodes the room as decimal ASCII; the engine and servicer hold it
as a Python int; the C++ scheduler never sees it), so this is a mechanical
type change with no logic change:

- proto (int32 -> int64): EncodeItemHandshake.bootstrap_room and
  DisaggregatedParams.bootstrap_room (tokenspeed_scheduler.proto),
  EncodeItemAssignment.bootstrap_room (tokenspeed_encoder.proto). Field
  numbers/names unchanged, so the change is wire-compatible for values that
  still fit i32; prost (build.rs) and the Python servicer stubs (setup.py,
  symlinked to the same canonical proto) both regenerate from it.
- Rust mints -> i64: the E->P mint (encode.rs) and the TokenSpeed P->D mint
  (helpers.rs). set_disaggregated's room param -> i64.
- The RDMA pixel-arena slot map keyed by the same room (rdma.rs): occupied
  DashMap, lease_and_write/free_room/reap_stale, and the free-notif tag parse
  all i32 -> i64.

The SGLang PD path (helpers.rs inject_sglang_bootstrap_metadata +
sglang_scheduler.proto) is deliberately left as int32 -- out of scope, and
widening it would touch the SGLang backend contract.

Source-only change: the gateway must be rebuilt (cargo, regenerates the prost
stubs) and the Python proto package reinstalled (regenerates the servicer
stubs) together, so a new gateway's wide rooms are never truncated by an old
int32 stub. Not compiled on this host (no Rust toolchain); cargo build/clippy
must run on a build node.

+++ model_gateway/src/routers/grpc/common/stages/encode.rs
@@ -122,7 +122,9 @@ impl PipelineStage for EncodeStage {
 
         for (global_index, split) in splits.into_iter().enumerate() {
             let worker = encode_pool[(rr_base + global_index) % pool_len].clone();
-            let bootstrap_room = rand::rng().random_range(0..i32::MAX);
+            // 63-bit room: no in-flight dedup, so a 2^31 space birthday-collides
+            // under load (silent embedding cross-wire). See the proto field doc.
+            let bootstrap_room = rand::rng().random_range(0..i64::MAX);
 
             // The handshake points the prefill at THIS worker's Mooncake bootstrap
             // endpoint (host/port), not its gRPC URL. One handshake per image: when
